Make verify_email an authenticated endpoint.
It was not previously, because it used to be that the user would
be emailed a link to click, which ideally would work whether or
not the user was logged in on the browser they clicked it on.
With the current method of using a token, this doesn't make much
sense. It is also a potential minor security risk - if a user
mistypes their email address, the owner of the mistyped address
is able to (before this change) verify the email address, which
in the future would potentially allow them to take control of
the account.